requirements.txt
truststore
httpx
fake-useragent
python-dotenv
tenacity
pydantic
pyyaml
